Create README.md
The README file for the published version of My Health Journal.  Primarily intended to error reports, requirements, comments, ideas, etc.

The data items will be stored here along the development status. There will be a hyperlink on the Home Page to a form to input the data. These data will be reported in a blog.